+++ R/namespace.R
@@ -1,2 +1,3 @@
 #' @import checkr
+#' @importFrom stats setNames
 NULL