+++ R/sample.R
@@ -1,0 +1,14 @@
+sample_population <- function(x, x2, n = 1) {
+  x <- x[.parameters]
+  x2 <- x2[.parameters]
+
+  stopifnot(identical(x[c("tmax", "Rt")], x2[c("tmax", "Rt")]))
+
+  x <- mapply(c, x[.parameters], x2[.parameters], SIMPLIFY = FALSE)
+  x <- lapply(x, function(y) stats::runif(n, min(y), max(y)))
+  x$FUN <- function(...) {list(...)}
+  x$SIMPLIFY <- FALSE
+  x <- do.call("mapply", x)
+  x <- lapply(x, FUN = function(y) {y[c("tmax", "Rt")] <- x2[c("tmax", "Rt")]; y})
+  x
+}