+++ R/yield.R
@@ -14,6 +14,7 @@ yield_mu <- function(mu, population, Ly, harvest, biomass) {
 #' @param biomass A flag indicating whether to calculate the yield in terms of the biomass versus number of individual fish.
 #' @param sanitize A flag indicating whether to replace negative or NaNs with 0.
 #' @param check A flag indicating whether to check the arguments.
+#' @return A double of the yield.
 #' @seealso \code{\link{ypr_population}}, \code{\link{ypr_optimize}},
 #' \code{\link{ypr_yields}} and \code{\link{ypr_plot}}
 #' @export