+++ tests/testthat/test-yield-limits.R
@@ -1,0 +1,10 @@
+context("yield-limits")
+
+test_that("yield_limits", {
+  population <- ypr_population()
+  population$Rk <- 10
+  yield <- ypr_yield_limits(population, ypr_population(), 3)
+
+  expect_identical(length(yield), 3L)
+  expect_identical(names(yield), c("estimate", "lower", "upper"))
+})