+++ NEWS.md
@@ -1,4 +1,6 @@
-- Removed a and fa as just scaling terms.
+- Added `ypr_yield_limits()` to account for uncertainty in parameters.
+- Increased upper limit for `Rk` from 10 to 100.
+- Removed `a` and `fa` as just scaling terms.
 
 # ypr 0.0.1